Phase 1 implementation, Bash completion
